backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
qdrant-client==1.7.0
sentence-transformers==2.3.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
openai==1.10.0
anthropic==0.18.1
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
numpy==1.26.3
pandas==2.1.4
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14